fix: Renamed a collection of components in advance of css refactor using css modules.
