+++ src/index.ts
@@ -15,13 +15,13 @@ export { default as ButtonFeatureIcon } from "./components/ButtonFeatureIcon.vue
 export { default as ButtonOutlineDark} from "./components/ButtonOutlineDark.vue"
 export { default as ButtonOutlineSmall} from "./components/ButtonOutlineSmall.vue"
 export { default as ButtonOutlineDarkSmall} from "./components/ButtonOutlineDarkSmall.vue"
-export { default as Card } from "./components/Card.vue"
+export { default as Card } from "./components/FishTankCard.vue"
 export { default as InputCheckbox } from "./components/FishTankCheckbox.vue"
 export { default as InputRadio } from "./components/FishTankRadio.vue"
 export { default as InputSwitch } from "./components/FishTankSwitch.vue"
-export { default as InputText } from "./components/InputText.vue"
-export { default as Modal } from "./components/Modal.vue"
-export { default as DialogBox } from "./components/DialogBox.vue"
+export { default as InputText } from "./components/FishTankTextInput.vue"
+export { default as Modal } from "./components/FishTankModal.vue"
+export { default as DialogBox } from "./components/FishTankDialogBox.vue"
 export { default as Tag } from "./components/FishTankTag.vue"
 
 //Todo - Remove exports above and retain only rebranded exports below
@@ -37,15 +37,14 @@ export { default as FishTankButtonDestructive } from "./components/ButtonDestruc
 export { default as FishTankButtonWhite } from "./components/ButtonWhite.vue"
 export { default as FishTankButtonOutline} from "./components/ButtonOutline.vue"
 export { default as FishTankButtonGradient } from "./components/ButtonGradient.vue"
-export { default as FishTankCard } from "./components/Card.vue"
+export { default as FishTankCard } from "./components/FishTankCard.vue"
 export { default as FishTankCheckbox } from "./components/FishTankCheckbox.vue"
 export { default as FishTankRadio } from "./components/FishTankRadio.vue"
 export { default as FishTankSwitch } from "./components/FishTankSwitch.vue"
-export { default as FishTankTextInput } from "./components/InputText.vue"
-export { default as FishTankModal } from "./components/Modal.vue"
-export { default as FishTankDialogBox } from "./components/DialogBox.vue"
+export { default as FishTankTextInput } from "./components/FishTankTextInput.vue"
+export { default as FishTankModal } from "./components/FishTankModal.vue"
+export { default as FishTankDialogBox } from "./components/FishTankDialogBox.vue"
 export { default as FishTankTag } from "./components/FishTankTag.vue"
-export { default as FishTankTagRemove } from "./components/InputTagRemove.vue"
 export { default as FishTankMultiSelectTag } from "./components/MultiSelectTag.vue"
 export { default as FishTankLoaderSpinner } from "./components/FishTankLoaderSpinner.vue"
 export { default as FishTankButtonGroup } from "./components/FishTankButtonGroup.vue"
